% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{bienvenue}
\alias{bienvenue}
\title{Fonction bonjour}
\usage{
bienvenue(jour)
}
\arguments{
\item{jour}{une chaine de carctere}
}
\value{
un print en console
}
\description{
Fonction qui renvoie un texte de
bienvenue pour un jour precis
}
\examples{
bienvenue("mardi")
\dontrun{
bienvenue("mercredi")
}

}
