% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{bienvenue_h}
\alias{bienvenue_h}
\title{Fonction bonjour heure}
\usage{
bienvenue_h(heure)
}
\arguments{
\item{heure}{une chaine de carctere}
}
\value{
un print en console
}
\description{
Fonction qui renvoie un texte de
bienvenue pour une heure precise
}
\examples{
bienvenue_h("09:00")
\dontrun{
bienvenue_h("10:00")
}

}
