% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{byebye}
\alias{byebye}
\title{Fonction byebye}
\usage{
byebye(nom)
}
\arguments{
\item{nom}{une chaine de carctere}
}
\value{
un print en console
}
\description{
Fonction qui renvoie au revoir
}
\examples{
byebye("Benoit")
\dontrun{
byebye("Boris")
}

}
