% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jour}
\alias{jour}
\title{Jour semaine}
\format{data.frame contenant 2 variables
\describe{
 \item{num}{un numeric correspondant au jour.}
 \item{nom}{une chaine de caratere correspondant au jour.}
 }}
\source{
INRA
}
\usage{
jour
}
\description{
Jour semaine
}
\examples{
 jour

}
\keyword{datasets}
