% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{monnouveaupackage}
\alias{monnouveaupackage}
\alias{monnouveaupackage-package}
\title{monnouveaupackage package}
\description{
Un package qui renvoie un message de
bienvenue. Sivous souhaitez plus d'onformation,
une vifgnette est disponible:\cr

\code{vignette(topic = "monnouveaupackage", package = "monnouveaupackage")}.\cr
Pour acceder aux donnees du rapport, rdv sur:  \cr
\code{system.file("brutes","cr_inra.docx",package="monnouveaupackage")}
}
